Add UT and clean code

+++ Takser/Infra/Services/TasksGroupService.cs
@@ -1,5 +1,8 @@
-﻿using System;
+﻿using Logger.Contracts;
+using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 using Takser.Infra.Options;
 using TaskData.Contracts;
@@ -13,11 +16,15 @@ namespace Tasker.Infra.Services
     public class TasksGroupService : ITasksGroupService
     {
         private readonly IDbRepository<ITasksGroup> mTasksGroupRepository;
+        private readonly ITasksGroupBuilder mTaskGroupBuilder;
         private readonly NameValidator mNameValidator;
+        private readonly ILogger mLogger;
 
-        public TasksGroupService(IDbRepository<ITasksGroup> TaskGroupRepository)
+        public TasksGroupService(IDbRepository<ITasksGroup> TaskGroupRepository, ITasksGroupBuilder tasksGroupBuilder, ILogger logger)
         {
             mTasksGroupRepository = TaskGroupRepository;
+            mTaskGroupBuilder = tasksGroupBuilder;
+            mLogger = logger;
             mNameValidator = new NameValidator(NameLengths.MaximalGroupNameLength);
         }
 
@@ -26,30 +33,31 @@ namespace Tasker.Infra.Services
             return await mTasksGroupRepository.ListAsync();
         }
 
-        public async Task<TasksGroupResponse> SaveAsync(ITasksGroup group)
+        public async Task<Response<ITasksGroup>> SaveAsync(string groupName)
         {
             try
             {
-                await mTasksGroupRepository.AddAsync(group);
+                ITasksGroup tasksGroup = mTaskGroupBuilder.Create(groupName, mLogger);
+                await mTasksGroupRepository.AddAsync(tasksGroup);
 
-                return new TasksGroupResponse(group);
+                return new Response<ITasksGroup>(tasksGroup, isSuccess: true);
             }
             catch (Exception ex)
             {
-                return new TasksGroupResponse($"An error occurred when saving tasks groups id {group.ID}: {ex.Message}");
+                return new Response<ITasksGroup>(isSuccess: false, $"An error occurred when saving tasks grou name {groupName}: {ex.Message}");
             }
         }
 
-        public async Task<TasksGroupResponse> UpdateAsync(string id, string newGroupName)
+        public async Task<Response<ITasksGroup>> UpdateAsync(string id, string newGroupName)
         {
             ITasksGroup groupToUpdate = 
-                await mTasksGroupRepository.FindByIdAsync(id);
+                await mTasksGroupRepository.FindAsync(id);
 
             if (groupToUpdate == null)
-                return new TasksGroupResponse("Group not found");
+                return new Response<ITasksGroup>(isSuccess: false, "Group not found");
 
             if (mNameValidator.IsNameValid(newGroupName))
-                return new TasksGroupResponse($"Group name '{newGroupName}' is invalid");
+                return new Response<ITasksGroup>(isSuccess: false, $"Group name '{newGroupName}' is invalid");
 
             groupToUpdate.Name = newGroupName;
 
@@ -57,29 +65,38 @@ namespace Tasker.Infra.Services
             {
                 await mTasksGroupRepository.UpdateAsync(groupToUpdate);
 
-                return new TasksGroupResponse(groupToUpdate);
+                return new Response<ITasksGroup>(groupToUpdate, isSuccess: true);
             }
             catch (Exception ex)
             {
-                return new TasksGroupResponse($"An error occurred when updating tasks group id {id}: {ex.Message}");
+                return new Response<ITasksGroup>(isSuccess: false, $"An error occurred when updating tasks group id {id}: {ex.Message}");
             }
         }
 
-        public async Task<TasksGroupResponse> RemoveAsync(string id)
+        public async Task<Response<ITasksGroup>> RemoveAsync(string id)
         {
-            ITasksGroup groupToRemove = await mTasksGroupRepository.FindByIdAsync(id);
+            ITasksGroup groupToRemove = await mTasksGroupRepository.FindAsync(id);
+
             if (groupToRemove == null)
-                return new TasksGroupResponse(groupToRemove, $"Entity group {id} not found. No deletion performed");
+                return new Response<ITasksGroup>(isSuccess: false, $"Entity group {id} not found. No deletion performed");
+
+            if (groupToRemove.Size > 0)
+            {
+                StringBuilder idsInGroup = new StringBuilder();
+                groupToRemove.GetAllTasks().Select(task => task.ID).ToList().ForEach(id => idsInGroup.Append($"{id}, "));
+                return new Response<ITasksGroup>(groupToRemove, isSuccess: false, $"Entity group {id} cannot be deleted. Please move or remove" +
+                        $"the next work tasks ids: {idsInGroup}");
+            }
 
             try
             {
                 await mTasksGroupRepository.RemoveAsync(groupToRemove);
 
-                return new TasksGroupResponse(groupToRemove);
+                return new Response<ITasksGroup>(groupToRemove, isSuccess: true);
             }
             catch (Exception ex)
             {
-                return new TasksGroupResponse($"An error occurred when removing tasks group id {id}: {ex.Message}");
+                return new Response<ITasksGroup>(isSuccess: false, $"An error occurred when removing tasks group id {id}: {ex.Message}");
             }
         }
     }
